ワーニングに対処．

+++ Graphics/ImageRGBATest.cpp
@@ -6,7 +6,7 @@ using namespace Crystal::Graphics;
 
 TEST( ImageRGBATest, ContructTest )
 {
-	ImageRGBA<unsigned int> image;
+	ImageRGBA<float> image;
 	EXPECT_EQ( 0, image.getWidth() );
 	EXPECT_EQ( 0, image.getHeight() );
 	//EXPECT_TRUE( image.getValues().empty() );
@@ -14,7 +14,7 @@ TEST( ImageRGBATest, ContructTest )
 
 TEST( ImageRGBATest, TestColor )
 {
-	ImageRGBA<unsigned int> color( 1, 1);
-	color.setColor(0, 0, ColorRGBA<unsigned int>::Red() );
-	EXPECT_EQ( ColorRGBA<unsigned int>::Red(), color.getColor(0, 0) );
+	ImageRGBA<float> color( 1, 1);
+	color.setColor(0, 0, ColorRGBA<float>::Red() );
+	EXPECT_EQ( ColorRGBA<float>::Red(), color.getColor(0, 0) );
 }
